+++ SwiftStudio/Assets.xcassets/Attachment.imageset/Contents.json
@@ -1,0 +1,23 @@
+{
+  "images" : [
+    {
+      "idiom" : "universal",
+      "filename" : "attachment [#1575]@1x.png",
+      "scale" : "1x"
+    },
+    {
+      "idiom" : "universal",
+      "filename" : "attachment [#1575].png",
+      "scale" : "2x"
+    },
+    {
+      "idiom" : "universal",
+      "filename" : "attachment [#1575]@3x.png",
+      "scale" : "3x"
+    }
+  ],
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  }
+}
