+++ SwiftStudio/BoardCreateViewController.swift
@@ -18,6 +18,7 @@ class BoardCreateViewController: UIViewController, UIImagePickerControllerDelega
     
     var boardData: BoardObject!
     var delegate: MainViewController?
+    var cellDelegate : BoardCell?
     
     private var ref : FIRDatabaseReference! = FIRDatabase.database().reference()
     private var boardRef : FIRDatabaseReference! = FIRDatabase.database().reference().child(boardPostChildName)
@@ -55,6 +56,12 @@ class BoardCreateViewController: UIViewController, UIImagePickerControllerDelega
         
         //네비게이션바 세팅
         navBar.items = [navigationItem]
+        
+        //2017.02.24
+        //수정할 값이 들어올경우 기본 텍스트에 표기
+        if boardData != nil {
+            self.textView.text = boardData.bodyText
+        }
     }
     
     
@@ -70,7 +77,9 @@ class BoardCreateViewController: UIViewController, UIImagePickerControllerDelega
         
         navigationItem.rightBarButtonItem?.isEnabled = false
         if textView.text.characters.count > 0 {
-            let value : [String : Any] = [
+            var key = boardRef.childByAutoId().key
+            
+            var value : [String : Any] = [
                 "text" : textView.text,
                 "recordTime" : FIRServerValue.timestamp(),
                 "author" : [
@@ -79,7 +88,30 @@ class BoardCreateViewController: UIViewController, UIImagePickerControllerDelega
                     "userName" : "song"
                 ]
             ]
-            boardRef.childByAutoId().setValue(value) { error, ref in
+            
+            //June Kang - 2017.02.23 추가적인 수정 로직 만듬
+            if(boardData != nil){
+                
+                key = boardData.boradKey!
+                value["editTime"] = FIRServerValue.timestamp()
+                value["recordTime"] = boardData.recordTimeDouble
+                
+                var userObj = value["author"] as! [String : Any]
+                userObj["uid"] = boardData.authorId
+                userObj["userName"] = boardData.authorName
+                userObj["profile_url"] = boardData.profileImgUrl
+                
+                value["author"] = userObj
+                
+                if boardData.like != nil{
+                    value["like"] = boardData.like
+                    value["likeCount"] = boardData.likeCount
+                }
+                
+            }
+            
+            //데이터가 있을경우 수정 없을경우 생성
+            boardRef.child(key).updateChildValues(value) { error, ref in
                 if let err = error {
                     debugPrint(err.localizedDescription)
                     Toast(text: "실패").show()
@@ -87,8 +119,14 @@ class BoardCreateViewController: UIViewController, UIImagePickerControllerDelega
                     self.textView.resignFirstResponder()
                     self.textView.text = nil
                     Toast(text: "성공").show()
+                    
+                    self.dismiss(animated: true, completion: {
+                        if self.boardData.like != nil{
+                            //수정후 처리
+                            self.delegate?.afterEdit(ref: ref, cell : self.cellDelegate!)
+                        }
+                    })
                 }
-                
                 self.navigationItem.rightBarButtonItem?.isEnabled = true
             }
         }
